refactor: Weather 아키텍처 구조 리팩토링
